+++ akshare/stock_feature/stock_a_indicator.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 """
-Date: 2023/4/6 20:20
+Date: 2024/02/14 01:24
 Desc: 市盈率, 市净率和股息率查询
 https://www.legulegu.com/stocklist
 https://www.legulegu.com/s/000001
@@ -22,8 +22,8 @@ def get_cookie_csrf(url: str = "") -> dict:
     :rtype: pandas.DataFrame
     """
     r = requests.get(url)
-    soup = BeautifulSoup(r.text, "lxml")
-    csrf_tag = soup.find("meta", attrs={"name": "_csrf"})
+    soup = BeautifulSoup(r.text, features="lxml")
+    csrf_tag = soup.find(name="meta", attrs={"name": "_csrf"})
     csrf_token = csrf_tag.attrs["content"]
     headers = {"X-CSRF-Token": csrf_token}
     return {"cookies": r.cookies, "headers": headers}
@@ -55,7 +55,7 @@ def stock_a_indicator_lg(symbol: str = "002174") -> pd.DataFrame:
     if symbol == "all":
         url = "https://legulegu.com/stocklist"
         r = requests.get(url)
-        soup = BeautifulSoup(r.text, "lxml")
+        soup = BeautifulSoup(r.text, features="lxml")
         node_list = soup.find_all(attrs={"class": "col-xs-6"})
         href_list = [item.find("a")["href"] for item in node_list]
         title_list = [item.find("a")["title"] for item in node_list]
@@ -81,14 +81,14 @@ def stock_a_indicator_lg(symbol: str = "002174") -> pd.DataFrame:
         )
         temp_df["trade_date"] = pd.to_datetime(temp_df["trade_date"]).dt.date
         temp_df[temp_df.columns[1:]] = temp_df[temp_df.columns[1:]].astype(float)
-        temp_df.sort_values(["trade_date"], inplace=True, ignore_index=True)
-        if len(set(temp_df['trade_date'])) < 10:
+        temp_df.sort_values(by=["trade_date"], inplace=True, ignore_index=True)
+        if len(set(temp_df['trade_date'])) <= 0:
             raise ValueError("数据获取失败, 请检查是否输入正确的股票代码")
         return temp_df
 
 
 def stock_hk_indicator_eniu(
-    symbol: str = "hk01093", indicator: str = "市盈率"
+        symbol: str = "hk01093", indicator: str = "市盈率"
 ) -> pd.DataFrame:
     """
     亿牛网-港股指标
@@ -116,7 +116,7 @@ def stock_hk_indicator_eniu(
         url = f"https://eniu.com/chart/dvh/{symbol}"
     elif indicator == "ROE":
         url = f"https://eniu.com/chart/roeh/{symbol}"
-    elif indicator == "市值":
+    else:
         url = f"https://eniu.com/chart/marketvalueh/{symbol}"
     r = requests.get(url)
     data_json = r.json()
@@ -128,7 +128,7 @@ if __name__ == "__main__":
     stock_a_indicator_lg_all_df = stock_a_indicator_lg(symbol="all")
     print(stock_a_indicator_lg_all_df)
 
-    stock_a_indicator_lg_df = stock_a_indicator_lg(symbol="501050")
+    stock_a_indicator_lg_df = stock_a_indicator_lg(symbol="301502")
     print(stock_a_indicator_lg_df)
 
     stock_hk_indicator_eniu_df = stock_hk_indicator_eniu(