+++ docs/changelog.md
@@ -71,6 +71,10 @@
 
 ## 更新说明详情
 
+1.12.56 fix: fix stock_a_indicator_lg interface
+
+    1. 修复 stock_a_indicator_lg 接口
+
 1.12.55 fix: fix stock_us_hist interface
 
     1. 修复 stock_us_hist 接口
@@ -3522,6 +3526,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.12.56 fix: fix stock_a_indicator_lg interface
 
 1.12.55 fix: fix stock_us_hist interface